feat: Reworked the entire action
BREAKING CHANGE: Uses .distinclude to find the files

+++ packer-action.bash
@@ -1,21 +1,50 @@
 #!/bin/bash
 set -e
 
-NEXT_VERSION=$(npm run semantic-release -- --ci=false --dry-run | grep -oP 'Published release \K.*? ')
-CURRENT_VERSION=$(< "$ACTION_VERSION_FILE" grep "Version" | head -1 | awk -F= "{ print $2 }" | sed 's/[* Version:,\",]//g' | tr -d ':space:')
+# Function to find plugin version in PHP files and return the current version
+find_plugin_version() {
+  local version
 
-sed -i "s/$CURRENT_VERSION/$NEXT_VERSION/g" "$ACTION_VERSION_FILE"
+  while IFS= read -r -d $'\0' file; do
+    if grep -q '^\s*\* Plugin Name:' "$file"; then
+      version=$(grep -oP '^\s*\*\s*Version:\s*\K.*' "$file" | tr -d '[:space:]')
+      break
+    fi
+  done < <(find . -maxdepth 1 -type f -name "*.php" -print0)
+
+  echo "$version"
+}
+
+# Function to read the files and folders from .distinclude
+read_distinclude() {
+  local distinclude_file=".distinclude"
+  while IFS= read -r line; do
+    if [ -n "$line" ]; then
+      cp -ar "$line" "/tmp/$ACTION_PLUGIN_SLUG"
+    fi
+  done < "$distinclude_file"
+}
+
+NEXT_VERSION=$(npm run semantic-release -- --ci=false --dry-run | grep -oP 'Published release \K.*? ')
+CURRENT_VERSION=$(find_plugin_version)
+
+if [ -z "$CURRENT_VERSION" ]; then
+  echo "Error: Current version not found in PHP files."
+  exit 1
+fi
+
+sed -i "s/$CURRENT_VERSION/$NEXT_VERSION/g" ./*.php
 
 rm -rf "/tmp/$ACTION_PLUGIN_SLUG" 2>/dev/null
 mkdir -p "/tmp/$ACTION_PLUGIN_SLUG"
 
-if [ -n "$ACTION_INCLUDE_FILES" ];
-then
-  echo "Copying files to plugin directory"
-  IFS=' ' read -r -a folders <<< "$ACTION_INCLUDE_FILES"
-
-  cp -ar ./*.php "/tmp/$ACTION_PLUGIN_SLUG"
-  cp -ar -- "${folders[@]}" "/tmp/$ACTION_PLUGIN_SLUG"
+# Read .distinclude and copy the files/folders if it exists
+if [ -f ".distinclude" ]; then
+  echo "Copying files from .distinclude"
+  read_distinclude
+else
+  echo "Copying all files"
+  cp -ar . "/tmp/$ACTION_PLUGIN_SLUG"
 fi
 
 cd /tmp
